[networking] Ignore stderr from ip netns for list of namespaces
The sosreport of a system with this issue:
```
    # ip netns
    Error: Peer netns reference is invalid.
    Error: Peer netns reference is invalid.
    test-ns
```

Shows that the networking plugin runs commands for the `Error:` lines:
(note the ebpf plugin does not due to differences addressed in PR #2306)
```
    # ./bin/sos report -o ebpf,networking --batch

    # tar tf /tmp/sosreport-*.tar.xz | grep ip_netns_exec
    .../sos_commands/ebpf/ip_netns_exec_test-ns_bpftool_net_list
    .../sos_commands/networking/ip_netns_exec_Error_ip6tables-save
    .../sos_commands/networking/ip_netns_exec_Error_ip6tables-save.1
    ...
    .../sos_commands/networking/ip_netns_exec_test-ns_ip6tables-save
    ...
```

This happens because the networking plugin calls `collect_cmd_output()`
with default `stderr=True` and does not handle such line type.

For the purposes of getting the list of network namespaces, it is OK to
ignore `stderr` since it does not provide that information. However, we
do want it in the archive, so it is fully documented for analysis/debug.

So change the call from `collect_cmd_output()` to `add_cmd_output()` to
include both `stdout` and `stderr` in the archive, and call `exec_cmd()`
that ignores `stderr` to get the list of network namespaces.

Note that the plugin _currently_ does not need `exec_cmd(stderr=False)`
to ignore `stderr`, as described in PR#2306, but will once/if that is
applied. However, with the next patch, `stderr=False` won't be needed.

Before:
```
  # tar tf /tmp/sosreport-*.tar.xz | grep ip_netns_exec
  .../sos_commands/ebpf/ip_netns_exec_test-ns_bpftool_net_list
  .../sos_commands/networking/ip_netns_exec_Error_ip6tables-save
  .../sos_commands/networking/ip_netns_exec_Error_ip6tables-save.1
  ...
  .../sos_commands/networking/ip_netns_exec_test-ns_ip6tables-save
  ...
```

After:
```
  # tar tf /tmp/sosreport-*.tar.xz | grep ip_netns_exec
  .../sos_commands/ebpf/ip_netns_exec_test-ns_bpftool_net_list
  .../sos_commands/networking/ip_netns_exec_test-ns_ip6tables-save
  .../sos_commands/networking/ip_netns_exec_test-ns_ip_-s_-s_neigh_show
  ...
```

And the `ip netns` contents with `stderr` still remain in the archive:
```
    # tar xf /tmp/sosreport-*.tar.xz --to-stdout \
      --wildcards '*/sos_commands/networking/ip_netns'
    Error: Peer netns reference is invalid.
    Error: Peer netns reference is invalid.
    test-ns
```

Test suite:

```
    # ./tests/simple.sh
    ...
    Everything worked!
```

Signed-off-by: Mauricio Faria de Oliveira <mfo@canonical.com>