[cleaner] Match MAC addrs in lines that end in punctuation
It was found that sometimes MAC addresses would not be obfuscated from
certain files when the MAC address was followed by punctuation, such as
periods. This was because the regex used to match IPv4 MAC addresses
needs to avoid matching a substring within IPv6 mac addresses.

Update the regex to compensate, and then override the `parse_line()`
method to pull the matched punctuation out of the matched MAC substring
so that obfuscation remains consistent.

Resolves: #2311

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>