[cleaner] Simplify MAC addr parsing from #2311
The previously merged #2311 was thought to be sufficient for remaining
cases where MAC addrs were followed by punctuation characters, however
it was found to be too restrictive in that it would leave quoted and
other forms of MAC addrs unobfuscated when those addrs were immediately
followed by other characters.

Simply the regex match to catch all of these, and further update the
`parse_lines()` override to properly trim down the match to just the
address substring.

Related: #2311
Related: #2305

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>