[redhat] Add podman to valid container env vars
In newer versions of podman, the `container` env var set inside the
container will be set to `podman` instead of `oci`. Add this value to
the list of valid values to use to determine if sos is running inside a
container or not, for when we want to capture from the host system.

Resolves: #2320

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>