[tests] Add test for newly loaded kernel modules
Adds a test to `simple.sh` to check if we loaded any new kernel modules
during a normal exeuction. This test will be somewhat limited in the
fact that this currently only tests the plugins that are loaded by
default on each tested distribution.

Also included is a minor typo fix for reporting where IP addresses were
not obfuscated, if any. This typo did not affect the actual test
however.

Resolves: #2326

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>