[wireless] Update enablement triggers and add predicate for iw commands
First, update the enablement trigger to use the `commands` tuple to use
a policy's defined PATH to search for the `iw` command, rather than
splitting out subclasses based on the `files` tuple.

Second, add a predicate that requires the `cfg80211` kmod to be
loaded before running any `iw` commands, as this command will load the
module if it is not already loaded.

Related: #2326

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>