[kubernetes] Fix CDK related file paths
The `files` property is used to determined if it's the kubernetes
master node is where the plugin is runnig since it's the only node
capable of running kubectl. In Ubuntu CDK environments the file that's
only available in the master is /root/cdk/cdk_addons_kubectl_config
while /root/cdk/kubeproxyconfig is also available in the workers.

Signed-off-by: Felipe Reyes <felipe.reyes@canonical.com>