[redhat] Automatically failover to dropbox when RHCP upload fails
When an upload to the RHCP fails for some reason, now automatically
failover to attempt an upload to the FTP dropbox.

This failover has been intended for some time, as we already failover
to the FTP location if we detect we're missing, for example, login
credentials for the Customer Portal. This change just extends that
intention to when we run into errors with the actual upload process.

Resolves: #2350

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>