[RpmPackageManager] Add discreet PackageManager subclass
Adds a discreet subclass of `PackageManager` for RPM based
distributions.

Note this also makes a minor alteration to the determination of relevant
commands based on the initialization of `PackageManager`, which will
make it easier to build out subclasses going forward, while still
allowing individual policies to override if desired.

Related: #2349

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>