[process] Drop -b from lsof calls, add timeout
The `-b` option to `lsof` is used to avoid kernel calls that have the
potential to block. However, this in turn strips useful information from
this collection when processes are using network resources like NFS
shares.

As we now have several layers of timeout control, it should be safe to
remove the `-b` option and instead apply a timeout to the
`add_cmd_output()` call to protect against a potential blocked call.

Closes: #2352
Resolves: #2670

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>