[s390] Collect dasd.conf file
From RHBZ:
The /etc/dasd.conf is the common configuration file for the dasd portion of s390,
similar to how /etc/multipath.conf is to multipath.  It has settings like
failover timings, readonly settings, diag settings, etc.  It's very useful to see
what the customer has set in this file when doing an initial analysis of s390 and
more specifically dasd (s390 storage) issues.  It's also useful to have when
suggesting dasd tuning changes to the customer.

Resolves: RHBZ#1919277

Signed-off-by: Jose Castillo <jcastillo@redhat.com>