[suse] Properly instantiate a package manager
The previous change to package managers removed SuSE's manual settings
without actually having it use the new `RpmPackageManager` abstraction.
This in turn made almost any sos execution fail on SuSE.

Fix that, by instantiating the package manager.

Credit to fluxcap1 on GitHub for identifying this flaw and the
resolution.

Related: #2389

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>