[sos] Align plugin options with destination names
This commit alters several option long-form names or destination names
to align those values in a sensible way. This serves to not only remove
some abiguity in option naming in code, but also to make it so that the
"effective options" line logged in every sos execution can be direction
copy-pasted as a working command.

Closes: #2288
Resolves: #2398

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>