[build] Add the new presets module to setup.py
When presets were split out from policies, they were not added to
`setup.py` in the list of packages. As such they would not be included
in any builds that relied upon using `setup.py` to generate a source
tarball.

Add those packages to the list so that they may be included.

Closes: #2390
Resolves: #2402

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>