[clean] Fix hostnames not being obfuscated when host plugin wasn't run
It was discovered that if a report was generated without the host plugin
being run, then the hostname parser would not obfuscate the localhost
hostname or domain at all. This was due to the fact that while the
default_mapping values were being read into the parser's map's dataset
correctly, they weren't being loaded into the 'intermediary' dicts that
the parser uses to separate short names from domain names.

Fix this by reading the contents of the map's dataset dict (which is
populated by the map file) into those intermediary dicts that the parser
uses.

Additionally, safeguard against overloading domain and host obfuscation
numbers when loading this way by setting the respective counts to the
next-highest number after the map loading is complete.

Resolves: #2406

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>