[clean] Allow specifying keywords in a text file
This commit adds the ability for users to provide a text file with a
list of newline-delimited keywords that should be obfuscated, rather
than requiring all keywords be specified either by the `--keywords`
option or configuration file settings.

Files may be provided via the new `--keyword-file` option which is
available to `clean`, `collect`, and `report`.

Closes: #2401
Resolves: #2408

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>