[hostname_map] Fix duplicating entries of domains
The previous commit bf622ae was to address an issue where the system's
hostname would not be obfuscated if the host plugin was not run for an
sos execution (provided that it was previously run at least once to
generate the initial obfuscation).

While this commit solved that particular problem, it injected the
domains into the "intermediary" dicts used by the mapping in a way that
meant after a certain number of further subsequent runs without the host
plugin, we'd end up duplicating our TLDs.

E.G. 'example.com' being obfuscated as 'obfuscatedomain0.com' would
eventually produce 'obfuscateddomain0.com.com'.

Correct this injection to ensure consistent obfuscation regardless of
the number of future executions without the host plugin.

Related: #2406

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>