[cleaner|hostname] Match uppercase domains for kerberos
Previously, kerberos realms would not be obfuscated due to the
convention that those are written in all uppercase. While the parser
would match these, the hostname parser would not obfuscate them due to
the use of intermediary dicts that were case sensitive.

Now, match this convention and if it is found, obfuscate as a whole
domain rather than a potential FQDN with a hostname.

Closes: #2254
Resolves: #2416

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>