[rhcos] update relevant packages and services
This replaces the `coreos-metadata` package with `afterburn`, which
has been in use for multiple releases now.

Additionally, this expands the list of services queried to include a
number of relevant Ignition services that are run early during boot.
Ignition related failures are responsible for a good portion of RHCOS
related problems, so grabbing those logs is prudent.

Signed-off-by: Micah Abbott <miabbott@redhat.com>