[tests] Add StageTwoReportTest as a base test class
This commits adds a `StageTwoReportTest` from which tests that need to
do "light" mocking should be defined. As the name implies, this is the
base class for our Stage 2 tests.

"Light" mocking for our purposes is currently defined as dropping files
in locations on the test system's filesystem, which may include
replacing binaries, and installing packages for the durtation of the
test case. At the end of the test, packages are uninstalled and mocked
files are removed. In the event that a file already exists at a given
location on the test system, that file is temporarily renamed with a
'.sostesting' extension. On test clean up, once mocked files are
removed, any files renamed previously are moved back to their original
location/name.

Mocking is controlled automatically during test setup, and is based upon
test class attributes, similar to how plugin enablement is dependent
upon plugin class attributes. For Stage 2 tests cases, this includes:

    	`files` - a list of file names. These names should represent the
		  "real" location on the filesystem where the mocked
		  files will be placed. The mocked content should be
		  placed under a similar file path under tests/test_data
		  E.G. using `files = ['/etc/foo/bar']` will place a
		  copy of `tests/test_data/etc/foo/bar` under
		  `/etc/foo/bar` on the test system.

    	`packages` - a dict of packages to install. The dict keys map to
		     the distribution the package names are used for,
		     e.g. `rhel` or `ubuntu`. The values are lists of
		     package names, optionally with version included.

This commit includes just the framework for these tests, and no actual
Stage 2 tests are included as of yet.

Related: #2431

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>