[sosnode] Update regex parsing for newer plugin option output
In order to determine the available plugins on each node, `collect`
parses the `--list` output. This relies on several multiline regexes
that are built from the headers from each section printed in that
output.

Since there are now options available to every plugin, the order of
output sections has changed, and `collect` needs to break parsing
earlier to avoid scraping non-plugin-name strings into the available
plugin list for each node.

Resolves: #2477

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>