[tests] Set local distro in StageOne to allow distro-specific tests
Moves the detection of the local system's distribution to StageOne, and
defines two constants to reference RH and Ubuntu distribution lists to
allow for distro-specific tests.

Further, make the expected_warnings_displayed test skipped if the local
test system is not a RH-family distro, as those warnings are not
expected on non RH-family systems.

Related: #2431

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>