[tests] Add a test for default expected plugin enablement
Adds a test case to ensure that the plugins we can expect to always be
enabled when a "normal" `sos report` command is run are in fact enabled.

First, test the distro-independent plugins that do not have specific
enablement triggers (and thus should always be enabled).

Second, have a distro-specific test for distro-specific plugins that are
also expected to always run on those systems.

Closes: #2365
Resolves: #2431

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>