[ubuntu|Policy] Fix exception when attempting upload
Fixes an issue where an upload attempt on Ubuntu systems would fail with
a traceback due to the upload archive's name not being available when we
first check to make sure we have an upload location, since the Ubuntu
default location requires an archive/file name.

Related to this, stop clobbering `upload_archive` and assign the archive
name to an `upload_archive_name` variable instead.

Closes: #2472
Resolves: #2479

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>