[process] Collect per-process files from /proc
Adds collection of per-process files from /proc/$pid, limited to a
number of pids controlled by the new `process.numprocs` option. This
option will default to the first 2048 processes. Using a value of `0`
for this option will result in all pids being collected.

In testing, the initial limit of 2048 process did not significantly
raise collection time or archive size. Note however, that collection of
20k or more processes does show a significant increase in plugin
execution time and archive size. This is however tempered, by the
whole-plugin timeout that sos imposes.

Related: #436
Closes: #542
Resolves: #1783
Resolves: #2498

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>