[tests] Correct test package installation
First, correct the logs plugin stagetwo test to specify the needed
package in a list, and second move the import to within the test class
that needs it so that we don't fail on the stageone tests if the package
is not present.

Second, correct the journal injections that this plugin test does to use
`pre_sos_setup()` instead of overriding `setup_mocking()`, and call the
former after the mocking has been setup.

Finally, add the resulting size of the collected journal if we fail the
size requirements of the test.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>