[devicemapper] Add check for dm_mod
It was discovered during testing that some distributions will have
`dm_mod` as an unloaded kmod rather than being a builtin. For the
moment, we cannot rely on checking loaded kernel mods as a predicate for
`dm_mod` as it does not get listed when it is a builtin.

Instead, check for the presence of `/sys/module/dm_mod` as the marker
for if `dm_mod` is either a builtin or a currently loaded module.

Related: #2499

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>