[tests] assertFileGlobInArchive should only fail if file is present
Like `assertFileCollected`, `assertFileGlobInArchive` should only fail
if the referenced collection is not in the archive *and* it is present
on the test system. If it is not present on the test system, sos cannot
collect it and it is thus not an error to be missing from the archive.

Related: #2499

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>