[policies] Improve kernel module loading to include builtins
Adds determination of builtin modules to extend the list of kernel
modules that are loaded via our parsing of `lsmod`.

This is first done by parsing `/usr/lib/modules/$(uname
-r)/modules.builtin`, and then by further inspecting
`/boot/config-$(uname -r)` for specific kconfig strings that are
well-known to build kernel modules into the running kernel, but for
those that do not appear in the modules.builtin file above.

Closes: #2500

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>