[report] Do not fail on skipping non-existing plugins
Specifying a non-existing plugin to skip is no longer considered a fatal
error. This was previously in place, in part, for third party utilities
calling sos to ensure adherence to known plugins. However, since
`collect` has been integrated into sos natively, this is far less of a
concern.

Specfying non-existing plugins will now only generate a warning message
for plugins wanting to be skipped (`-n`). If non-existing plugins are
specified for enablement via `-e` or `-o`, that will still generate a
fatal error.

Closes: #537
Closes: #1723
Resolves: #2517

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>