[tests] Add test class for when exceptions are expected in an execution
Adds a new `StageOneReportExceptionTest` class that is to be used for
tests that should ensure an exception IS hit, for example when testing
plugin enablement or invalid options.

Included with this commit is a small set of plugin enablement tests.
Future tests for component and plugin specific exception handling should
be built using this class and design approach.

Resolves: #2518

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>