[cleaner] Remove binary files by default
Binary files generally speaking cannot be obfuscated, and as such we
should remove them from archives being obfuscated by default so that
sensitive data is not mistakenly included in an obfuscated archive.

This commits adds a new `--keep-binary-files` option that if used will
keep any encountered binary files in the final archive. The default
option of `false` will ensure that encountered binary files are removed.

The number of removed binary files per archive is reported when
obfuscation is completed for that archive.

Closes: #2478
Resolves: #2524

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>